+++ facebook_auth_platform_interface/lib/src/access_token.dart
@@ -1,42 +1,42 @@
 /// Class that contains the facebook access token data
 class AccessToken {
   /// DateTime with the expires date of this token
-  final DateTime? expires;
+  final DateTime expires;
 
   /// DateTime with the last refresh date of this token
-  final DateTime? lastRefresh;
+  final DateTime lastRefresh;
 
   /// the facebook user id
-  final String? userId;
+  final String userId;
 
   /// token provided by facebook to make api calls to the GRAPH API
-  final String? token;
+  final String token;
 
   // the facebook application Id
-  final String? applicationId;
+  final String applicationId;
 
   final String? graphDomain;
 
   /// list of string with the rejected permission by the user
-  final List<String>? declinedPermissions;
+  final List<String> declinedPermissions;
 
   /// list of string with the approved permission by the user
-  final List<String>? grantedPermissions;
+  final List<String> grantedPermissions;
 
   // is `true` when the token is expired
-  final bool? isExpired;
+  final bool isExpired;
 
   /// constrcutor
   AccessToken({
-    this.declinedPermissions,
-    this.grantedPermissions,
-    this.userId,
-    this.expires,
-    this.lastRefresh,
-    this.token,
-    this.applicationId,
+    required this.declinedPermissions,
+    required this.grantedPermissions,
+    required this.userId,
+    required this.expires,
+    required this.lastRefresh,
+    required this.token,
+    required this.applicationId,
     this.graphDomain,
-    this.isExpired,
+    required this.isExpired,
   });
 
   /// convert the data provided for the platform channel to one instance of AccessToken
@@ -64,8 +64,8 @@ class AccessToken {
   Map<String, dynamic> toJson() => {
         'userId': userId,
         'token': token,
-        'expires': expires?.toIso8601String(),
-        'lastRefresh': lastRefresh?.toIso8601String(),
+        'expires': expires.toIso8601String(),
+        'lastRefresh': lastRefresh.toIso8601String(),
         'applicationId': applicationId,
         'graphDomain': graphDomain,
         'isExpired': isExpired,