+++ facebook_auth_web/CHANGELOG.md
@@ -1,3 +1,5 @@
+### 2.1.1
+- Updated flutter_facebook_auth_platform_interface: ^2.1.1
 ### 2.1.0-dev.0
 - Added Unit test.
 - Removed MethodChannel.